Move add_blob to test utils
