+++ aquadoggo/src/test_utils/mod.rs
@@ -12,7 +12,7 @@ pub use config::TestConfiguration;
 pub use db::{drop_database, initialize_db, initialize_sqlite_db};
 pub use helpers::{build_document, doggo_fields, doggo_schema, schema_from_fields};
 pub use node::{
-    add_document, add_schema, add_schema_and_documents, populate_and_materialize,
+    add_blob, add_document, add_schema, add_schema_and_documents, populate_and_materialize,
     populate_store_config, update_document, TestNode,
 };
 pub use runner::{test_runner, test_runner_with_manager, TestNodeManager};
